+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/docs
 
+## 1.3.0
+
+### Minor Changes
+
+- Add users to organizations
+
 ## 1.2.0
 
 ### Minor Changes